Definitions are reset to "original" definition when multiple definitions of a data type.
IE. a new definition based on Umbraco.DropDown gets reset to original "DropDown" (-42).
Happens in 249: dataTypeDefintion = dataTypeDefinitions.First();
With the "new" types, legacyeditorid == Guid.Empty, so if predicate on 228 resolves to true. _dataTypeService.GetDataTypeDefinitionByPropertyEditorAlias(propertyEditorAlias) returns null, so the fresh nice correct datatype already fetched is tossed away and replaced with the built-in.

Should probably not pull this, but turn the entire if around. If definitionId == definition.Key && propertyEditorAlias == definition.PropertyEditorAlias, skip trying to fix it. It's all good. :)