Content awareness (v1)
Added Content Awareness to DataType saving, principly when you pick a
content node in something like a MNTP the id is stored in the settings.
this version of uSync tries to turn those ids into paths, and then turn
the paths into IDs making the content ID portable. (in theory)