USync issue with Doc Types enabled as List View
Versions- Umbraco 7.2.1 / USync 2.2.
I found this fix-
https://github.com/KevinJump/jumps.umbraco.usync/commit/70ec3f9fd6280b9037e607f17066e88e0a0d634c
which met with my conclusion that the pre-values were being deleted and
recreated every time USync ran in to the database, explaining why the
content values appeared to drop (although I could see they were still
there in the database, but they referenced prevalue IDs that no longer
existed.
So this problem does seem to have been fixed by upgrade to USync 2.4.4
to take this fix (yay!), but I now seem to have introduced this (for me
far more troublesome) issue of the Data types on the Document Types
changing away from the custom data type to the property editor. For me
this breaks stuff in a far more extreme way. It seems that because
GetDataTypeDefinitionByPropertyEditorAlias simply finds the first item
when there could be more than one as Alias are not unique. The fix is to
try and use the old method, then fall back onto the new method but with
the new method try and match on the guid to find the correct item,  then
finally simply pick the first or default.
Background on the issue See:
http://our.umbraco.org/projects/developer-tools/usync/usync/60391-Document-type-properties-for-custom-data-types-issue-(u721)

http://our.umbraco.org/projects/developer-tools/usync/usync/60391-Document-type-properties-for-custom-data-types-issue-(u721)